wav_file_decoder: scale waveform to screen width
Scales the waveform output to the size of the window generated by the
sdl2 canvas. Second displays the waveform as a set of lines that spread
from the center of the screen rather than singular points on a graph.
Also adds in second markers when displaying the waveform os it is more
clear how long the audio file would play for.

TODO: begin MP3 compression research and add in ability to play the
analyzed file directly from the window, maybe on pressing [space]

Signed-off-by: Ryan Sullivan <ryanbakersullivan@gmail.com>